update the ocaml configuration script with the perl one
we have a fork of Zarith in MirageOS that builds with dune, in
order to do cross-compilation and embedded builds more easily.
As part of that port, we use the ocaml script to configure the
repository instead of the perl one (to remove the need for perl
to be installed).

this PR syncs the functionality of the perl script `z_pp.pl`
with the ocaml version, so that the configurations from both
are identical.  The arg interface to the ocaml script changes
slightly to account for the OCaml compiler version options.
This script doesnt seem to be currently used, but I figured it
would be useful to send this upstream anyway in case there's
interest.