Fix parallel build (#93)
When running "make -j", the build fails because there are no rules to
build .cmti files. (They are byproducts of .cmi files.)

It also fails because it invokes ocamlmklib twice for generating both
zarith.cmxa and zarith.a, thus corrupting files.

Thus, instead of depending on the whole content of TOINSTALL, the "all"
rule now depends only on a subset large enough to trigger all the rules.

The downside is that one can no longer type "make zarith.a". Instead, one
has to type "make zarith.cmxa" (or just "make").